Move separated chart widgets to dedicated files
